requirements-dev.txt
coverage==4.4.1
idna==2.5
olefile==0.44
py==1.4.34
pytest==3.1.2
pytest-cov==2.5.1
pytest-runner==2.11.1
requests-mock==1.3.0
requests-toolbelt==0.8.0
six==1.10.0
tox==2.7.0
lorem==0.1.1